+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [📜 Changelog](#-changelog)
 * [📖 Table of contents](#-table-of-contents)
 * [0.x.y](#0xy)
+  * [0.0.5](#005)
   * [0.0.4](#004)
   * [0.0.3](#003)
   * [0.0.2](#002)
@@ -13,6 +14,10 @@
 <!-- TOC -->
 
 # 0.x.y
+
+## 0.0.5
+- Changed create typed axios function to emit verbs that now reject problem details.
+- For handling problem details in catch clauses, the create typed axios function now returns two utility functions too. They allow handling problem details / creating problem detail handler.
 
 ## 0.0.4
 - Changed export structure.